// imports/source/AplikasiBian-main/frontend/src/lib/types.ts
export interface Expense {
  id: string;
  date: string;
  amount: number;
  category: string;
  description: string;
}